enchantjs-fix を enchantjs/fix に変更
